Camera restraints (buggy)
The camera is now restrained to a play area set by a BoxCollider2D on
the GameManager object. It works correctly on the Y axis, but on the X
axis, the camera is restrained beyond the play area's boundaries, which
is a bug.  It seems to be some kind of offset bug